pbr: Add istio SNAT IP in LRP for host access
Istio service mesh historically needed a sidecar container
in every pod that participated in the service mesh. Istio
GA'd a new mode called "ambient" last year wherein there
would be a single proxy per node.

Without this patch, when istio is installed in ambient mode
alongside multus and ovn-kubernetes CNI plugin, pod health
checks fail.

Istio in ambient mode adds SNAT rules in the node wherein it
SNATs packets to 169.254.7.127 when kubelet health checks
the pods. The packet enters OVN via ovn-k8s-mp0

When the pod responds, the packet will go to the
cluster router. We need a logical router policy
to send the packet back to the switch with
ovn-k8s-mp0 as the next hop

Signed-off-by: Guru Shetty <gurus@nvidia.com>